requirements.txt
openai
elevenlabs
pyaudio
websocket-client
pygame
numpy
